requirements.txt
cryptography==42.0.8
fastapi==0.111.0
Pillow==10.4.0
pydantic==2.8.2
Requests==2.32.3
SQLAlchemy==2.0.31
uvicorn
python-multipart